added middleware for purchase

+++ routes/group/v1/purchase.php
@@ -6,7 +6,7 @@ use App\Http\Controllers\LoginController;
 use App\Http\Controllers\Api\v1\PurchaseController;
 
 Route::controller(PurchaseController::class)->prefix('purchase/v1')->group(function () {
-    Route::post('/','add')->middleware('artists.accessor')->name('purchases.add');
-    Route::get('{purchase}','list')->name('purchases.list');
-    Route::delete('{purchase}', 'destroy')->name('purchases.destroy');
+    Route::post('/','add')->middleware('auth:sanctum')->name('purchases.add');
+    Route::get('{purchase}','list')->middleware('auth:sanctum')->name('purchases.list');
+    Route::delete('{purchase}', 'destroy')->middleware('admin')->name('purchases.destroy');
 });
